+++ api/app/crud/location.py
@@ -4,12 +4,8 @@ from typing import List
 from bson import ObjectId
 
 from ..db import Database
-from ..models.location import (
-    GeoJSONPolygon,
-    LocationInputCreate,
-    LocationInputDatabase,
-    LocationOutputDatabase,
-)
+from ..models.location import (GeoJSONPolygon, LocationInputCreate,
+                               LocationInputDatabase, LocationOutputDatabase)
 from .errors import EntryNotFound
 
 
@@ -47,16 +43,16 @@ async def create_location(
     """Create a new location"""
     # reformat the data into the database object
     loc_db_fmt = LocationInputDatabase(
-        display_name=location.display_name, location={**location.dict()}
+        display_name=location.display_name, location={**location.model_dump()}
     )
     # store data in database
-    doc = await db.location_collection.insert_one(loc_db_fmt.dict())
+    doc = await db.location_collection.insert_one(loc_db_fmt.model_dump())
     # create data representation of inserted object
     inserted_id = doc.inserted_id
     return LocationOutputDatabase(
         id=str(inserted_id),
         created_at=ObjectId(inserted_id).generation_time,
-        **loc_db_fmt.dict(),
+        **loc_db_fmt.model_dump(),
     )
 
 
@@ -65,11 +61,15 @@ async def get_locations_within_bbox(
 ) -> List[LocationOutputDatabase]:
     """Query database for locations that are within a given bbox."""
     # read field names from data model
-    fields = LocationOutputDatabase.__fields__
+    fields = LocationOutputDatabase.model_fields
     param_location = fields["location"]
     # query database
     cursor = db.location_collection.find(
-        {f"{param_location}.coordinates": {"$geoWithin": {"$geometry": bbox.dict()}}}
+        {
+            f"{param_location}.coordinates": {
+                "$geoWithin": {"$geometry": bbox.model_dump()}
+            }
+        }
     )
     # convert locations to db representations object
     return [_document_to_db_obj(loc) for loc in await cursor.to_list(None)]