+++ api/app/crud/tags.py
@@ -1,14 +1,14 @@
 """Functions for computing tags."""
-from ..models.phenotype import ElementType, ElementTypeResult
+import logging
+
+from prp.models.phenotype import ElementType, ElementTypeResult
+from prp.models.typing import TypingMethod
+
 from ..models.sample import SampleInDatabase
-from ..models.tags import (
-    ResistanceTag,
-    Tag,
-    TagList,
-    TagSeverity,
-    TagType,
-    VirulenceTag,
-)
+from ..models.tags import (ResistanceTag, Tag, TagList, TagSeverity, TagType,
+                           VirulenceTag)
+
+LOG = logging.getLogger(__name__)
 
 
 # Phenotypic tags
@@ -90,7 +90,25 @@ def add_mrsa(tags: TagList, sample: SampleInDatabase) -> Tag:
     tags.append(tag)
 
 
-ALL_TAG_FUNCS = [add_pvl, add_mrsa]
+def add_sxt_type(tags: TagList, sample: SampleInDatabase) -> Tag:
+    """Check if sample SXT type."""
+    for type_res in sample.typing_result:
+        if type_res.type == TypingMethod.STX.value:
+            tag = Tag(
+                type=TagType.TYPING,
+                label=type_res.result.gene_symbol.upper(),
+                description="",
+                severity=TagSeverity.INFO,
+            )
+            tags.append(tag)
+
+
+# Tagging functions with the species they are applicable for
+ALL_TAG_FUNCS = [
+    {"species": ["Staphylococcus aureus"], "func": add_pvl},
+    {"species": ["Staphylococcus aureus"], "func": add_mrsa},
+    {"species": ["Escherichia coli"], "func": add_sxt_type},
+]
 
 
 def compute_phenotype_tags(sample: SampleInDatabase) -> TagList:
@@ -98,5 +116,8 @@ def compute_phenotype_tags(sample: SampleInDatabase) -> TagList:
     tags = []
     # iterate over tag functions to build up list of tags
     for tag_func in ALL_TAG_FUNCS:
-        tag_func(tags, sample)
+        major_spp = sample.species_prediction[0].scientific_name
+        LOG.debug("Major spp %s in %s", major_spp, str(tag_func["species"]))
+        if major_spp in tag_func["species"]:
+            tag_func["func"](tags, sample)
     return tags