+++ frontend/app/blueprints/api/views.py
@@ -2,7 +2,8 @@
 import json
 
 import requests
-from app.bonsai import TokenObject, add_samples_to_basket, remove_samples_from_basket
+from app.bonsai import (TokenObject, add_samples_to_basket,
+                        remove_samples_from_basket)
 from app.models import SampleBasketObject
 from flask import Blueprint, flash, jsonify, request
 from flask_login import current_user, login_required