+++ CHANGELOG.md
@@ -7,6 +7,8 @@
  - User can choose clustering method from the sample basket
  - Added spp prediction result to sample view
  - Find similar samples default to 50
+ - Added STX typing for _Escherichia coli_ samples
+ - CLI command for updating tags for all samples
 
 ### Fixed
 
@@ -18,3 +20,4 @@
  - Renamed cgvis to Bonsai
  - Complete rewrite of the project
  - Removed N novel cgmlst alleles from cgmlst qc card
+ - Use data models from JASEN Pipleine Result Processing application